single creature updated stubbed other types
