--- src/components/database/cards.js
@@ -1,35 +1,0 @@
-export default class CardsDB {
-    constructor(API) {
-        // this.api = API;
-        // this.attacks = db.addCollection('attacks');
-        // this.battlegear = db.addCollection('battlegear');
-        // this.creatures = db.addCollection('creatures');
-        // this.locations = db.addCollection('locations');
-        // this.mugic = db.addCollection('mugic');
-        // this.db = db;
-    }
-
-    setupCreatures(spreadsheet) {
-      this.getSpreadsheet(spreadsheet, function(json) {
-
-      });
-    }
-
-    setupMugic(spreadsheet) {
-      this.getSpreadsheet(spreadsheet, function(json) {
-
-      });
-    }
-
-    setupBattleGear(spreadsheet) {
-      this.getSpreadsheet(spreadsheet, function(json) {
-
-      });
-    }
-
-    setupLocations(spreadsheet) {
-      this.getSpreadsheet(spreadsheet, function(json) {
-
-      });
-    }
-}
