+++ src/components/portal/Category/Creatures.js
@@ -4,54 +4,20 @@ import { Link } from 'react-router';
 import PageNotFound from '../../PageNotFound';
 import API from '../../SpreadsheetData';
 import s from '../../../styles/app.style';
+import UnderConstruction from '../../UnderConstruction';
 
 export default class Creatures extends React.Component {
 
-  constructor(props) {
-    super (props);
-    this.state = {tribe: '', creatures: {}};
-  }
-
-  componentWillReceiveProps(nextProps) {
-    this.getData(nextProps);
-  }
-
-  componentDidMount() {
-    this.getData(this.props);
-  }
-
-  // ** Process the tribe ** //
-  // /portal/Creatures/
-  // /portal/{Tribe}/Creatures/
-  // The first / gets counted
-  getData(props) {
-    if (props.children) return this.props = props;
-    let path = props.location.pathname.split("/");
-    if (path[path.length-1] == "") path.pop(); // Remove trailing backslash
-
-    // Set tribe
-    let tribe = (path.length === 4) ? path[2] : "All";
-    this.setState({tribe: tribe});
-
-    // For each tribe, get its spreadsheet, set the state
-    var self = this;
-    let urls = (tribe == "All") ? API.Creatures : {[tribe]: API.Creatures[tribe]};
-    Object.keys(urls).map((tribe) => {
-      API.getSpreadsheet(urls[tribe], (data) => {
-        self.setState({creatures:
-          Object.assign(self.state.creatures, {[tribe]: data})
-        });
-      });
-    });
-    // self.setState({creatures: this.state.creatures.concat([data])});
-  }
-
-  hacks(event) {
-    console.log(event);
-    this.setState({click: true});
-  }
-
   render() {
+    if (this.props.children) {
+      return (<div>{this.props.children}</div>);
+    }
+    return (
+      <UnderConstruction location={this.props.location}/>
+    );
+  }
+
+  fakerender() {
     if (this.props.children) {
       return (<div>{this.props.children}</div>);
     }
