+++ src/components/portal/Single/Creature.js
@@ -1,9 +1,7 @@
 import React from 'react';
 import Interactive from 'react-interactive';
 import { Link } from 'react-router';
-// import {browserHistory} from 'react-router';
 import PageNotFound from '../../PageNotFound';
-import UnderConstruction from '../../UnderConstruction';
 import API from '../../SpreadsheetData';
 import s from '../../../styles/app.style';
 import {observer, inject} from 'mobx-react';
@@ -15,37 +13,6 @@ export default class SingleCreature extends React.Component {
   // /portal/Creatures/{Tribe}/{Name}
   // /portal/{Tribe}/Creatures/{Name}
   // The first / gets counted
-  getData(props) {
-
-
-    var name = decodeURIComponent(path[4]);
-
-    var self = this;
-    var creature = API.portal.creatures.findOne(name);
-    if (creature) self.setState({"creature": creature});
-    else self.setState({creature: "n/a"});
-    console.log(creature);
-    API.getSpreadsheet(API.Creatures[tribe], (data) => {
-      data.map((item, i) => {
-        if (item.title.$t == name)
-          self.setState({creature: item });
-      });
-      // If no creature set as false
-      if (!self.state.creature) {
-        self.setState({creature: "n/a"});
-      }
-    });
-    API.getSpreadsheet(API.Creatures_Card_Data, (data) => {
-      data.map((item, i) => {
-        if (item.title.$t == name) self.setState({card_data: item });
-      });
-      // If no card_data set as false
-      if (!self.state.card_data) {
-        self.setState({card_data: "n/a"});
-      }
-    });
-  }
-
   render() {
     const store = API;
 
@@ -74,14 +41,18 @@ export default class SingleCreature extends React.Component {
     //   return (<span>Invalid Tribe: {tribe}</span>);
     // }
 
+    if (!store.cards.built.includes("creatures_Cards")) {
+      store.cards.setupCreatures("Cards");
+      return (<span>Loading...</span>);
+    }
+
     if (!store.portal.built.includes("creatures_"+tribe)) {
       store.portal.setupCreatures(tribe);
       return (<span>Loading...</span>);
     }
-    // TODO load card data
 
     const creature = store.portal.creatures.findOne({'gsx$name': path[4]});
-    console.log(creature);
+    const card_data = store.cards.creatures.findOne({'gsx$name': path[4]});
     if (!creature) {
       return(<PageNotFound location={this.props.location}/>);
     }
@@ -94,9 +65,12 @@ export default class SingleCreature extends React.Component {
       return <p key={i}><Interactive as={Link} {...s.link} to={"/portal/Battlegear/"+item}><span>{item}</span></Interactive></p>;
     });
 
+    const elements = card_data.gsx$elements.split(/[ ,]+/).map((item, i) => {
+      return <img className="icon" src={"/src/img/icons/elements/"+item.toLowerCase()+".png"} alt={item} key={i}></img>;
+    });
+
     return (
       <div className={"creature " + tribe.toLowerCase()}>
-        <UnderConstruction location={this.props.location}/>
         <h1>{creature.gsx$name}</h1>
         <img className="splash" src={store.base_image + creature.gsx$splash}></img>
         <hr />
@@ -139,113 +113,60 @@ export default class SingleCreature extends React.Component {
           <strong>Weight (lb):</strong><br />
           {creature.gsx$weight}
         </div>
-      </div>
-    );
-  }
-
-  fakerender() {
-    var self = this;
-
-    const elements = card_data.gsx$elements.$t.split(/[ ,]+/).map((item, i) => {
-      return <img className="icon" src={"/src/img/icons/elements/"+item.toLowerCase()+".png"} alt={item} key={i}></img>;
-    });
-
-    const locations = creature.gsx$location.$t.split(/[,]+\s*/).map((item, i) => {
-      return <p key={i}><Interactive as={Link} {...s.link} to={"/portal/Locations/"+item}><span>{item}</span></Interactive></p>;
-    });
-
-    const battlegear = creature.gsx$battlegear.$t.split(/[,]+\s*/).map((item, i) => {
-      return <p key={i}><Interactive as={Link} {...s.link} to={"/portal/Battlegear/"+item}><span>{item}</span></Interactive></p>;
-    });
-
-    return(
-      <div className={"creature " + this.state.tribe.toLowerCase()}>
-        <h1>{creature.gsx$name.$t}</h1>
-        <img className="splash" src={API.base_image + creature.gsx$splash.$t}></img>
-        <hr />
-        <div>
-          <strong>Appearance:</strong><br />
-          {creature.gsx$appearance.$t}
-        </div>
-        <hr />
-        <div>
-          <strong>Background:</strong><br />
-          {creature.gsx$background.$t}
-        </div>
-        <hr />
-        <div>
-          <strong>Details:</strong><br />
-          {creature.gsx$details.$t}
-        </div>
-        <hr />
-        <div>
-          <strong>Favorite Battlegear(s):</strong><br />
-          {battlegear}
-        </div>
-        <hr />
-        <div>
-          <strong>Favorite Location(s):</strong><br />
-          {locations}
-        </div>
-        <hr />
-        <div>
-          <strong>Height (ft):</strong><br />
-          {creature.gsx$height.$t}
-        </div>
         <hr />
         <div>
           <strong>Special Abilities:</strong><br />
-          {creature.gsx$specialabilities.$t}
+          {creature.gsx$specialabilities}
         </div>
         <hr />
         <div>
           <strong>Weight (lb):</strong><br />
-          {creature.gsx$weight.$t}
+          {creature.gsx$weight}
         </div>
         <hr />
         <div>
           <strong>Card ID: </strong>
-          {card_data.gsx$cardid.$t}
+          {card_data.gsx$cardid}
         </div>
         <hr />
         <div>
           <strong>Set: </strong>
-          {card_data.gsx$set.$t}
+          {card_data.gsx$set}
         </div>
         <hr />
         <div>
           <strong>Rarity: </strong>
-          {card_data.gsx$rarity.$t}
+          {card_data.gsx$rarity}
         </div>
         <hr />
         <div>
-          <strong>Tribe: </strong>{this.state.tribe}
-          <img className="icon" src={"/src/img/icons/tribes/"+this.state.tribe.toLowerCase()+".png"}></img>
+          <strong>Tribe: </strong>{tribe}
+          <img className="icon" src={"/src/img/icons/tribes/"+tribe.toLowerCase()+".png"}></img>
         </div>
         <hr />
         <div>
           <strong>Ability:</strong><br />
-          {card_data.gsx$ability.$t}
+          {card_data.gsx$ability}
         </div>
         <hr />
         <div>
           <strong>Courage: </strong>
-          {card_data.gsx$courage.$t}
+          {card_data.gsx$courage}
         </div>
         <hr />
         <div>
           <strong>Power: </strong>
-          {card_data.gsx$power.$t}
+          {card_data.gsx$power}
         </div>
         <hr />
         <div>
           <strong>Speed: </strong>
-          {card_data.gsx$speed.$t}
+          {card_data.gsx$speed}
         </div>
         <hr />
         <div>
           <strong>Wisdom: </strong>
-          {card_data.gsx$wisdom.$t}
+          {card_data.gsx$wisdom}
         </div>
         <hr />
         <div>
@@ -254,17 +175,17 @@ export default class SingleCreature extends React.Component {
         <hr />
         <div>
           <strong>Energy: </strong>
-          {card_data.gsx$energy.$t}
+          {card_data.gsx$energy}
         </div>
         <hr />
         <div>
           <strong>Flavortext:</strong><br />
-          {card_data.gsx$flavortext.$t}
+          {card_data.gsx$flavortext}
         </div>
         <hr />
         <div>
           <strong>Mugic Ability: </strong>
-          {card_data.gsx$mugicability.$t}
+          {card_data.gsx$mugicability}
         </div>
       </div>
     );
